+++ modules/sample/src/test/js/client/pass/SyncPassTest.js
@@ -1,5 +1,5 @@
-import { UnitTest, assert } from '@ephox/bedrock-client';
+import { Assert, UnitTest } from '@ephox/bedrock-client';
 
 UnitTest.test('SyncPass Test', function () {
-  assert.eq(1, 1);
+  Assert.eq('', 1, 1);
 });